Tested and fixed multiple bugs on the implementation of the algorithms and in the UI. Improved the UI. Added more parameter on ScriptableObjects to be configured easily.

+++ MatchmakingServer/Assets/Core/Scripts/ScriptableObjects/EloRatingParameters.cs
@@ -5,7 +5,7 @@ namespace MM.Server.Core
     [CreateAssetMenu(fileName = "EloRatingParameters", menuName = "ScriptableObjects/EloRatingParameters", order = 1)]
     public class EloRatingParameters : ScriptableObject
     {
-        public int ScaleFactor = 400;
-        public int UpdateSpeed = 32;
+        public float ScaleFactor = 400;
+        public float UpdateSpeed = 32;
     }    
 }
